Add missing row and align buttons properly
Due to tunnel vision after working too long, an entire row of
buttons was forgotten in the initial layout commit However, due to
well planned structuring of elements, adding the missing row was a
trivial change that did not break any of the css formatting or
alignment of other buttons. Taking a break also helped me come back
and fix the alignment of the bottom most row.

Taking a step back is an important lesson that should not be
forgotten.